Dimmer: set semi-period on uint16_t

+++ lib/MycilaDimmer/MycilaDimmer.cpp
@@ -60,12 +60,12 @@ static const uint32_t DIMMER_MAX = (1 << MYCILA_DIMMER_RESOLUTION) - 1;
 static const uint32_t TABLE_PHASE_SCALE = (TABLE_PHASE_LEN - 1U) * (1UL << (16 - MYCILA_DIMMER_RESOLUTION));
 static const uint16_t TABLE_PHASE_DELAY[TABLE_PHASE_LEN] PROGMEM = {0xefea, 0xdfd4, 0xd735, 0xd10d, 0xcc12, 0xc7cc, 0xc403, 0xc094, 0xbd6a, 0xba78, 0xb7b2, 0xb512, 0xb291, 0xb02b, 0xaddc, 0xaba2, 0xa97a, 0xa762, 0xa557, 0xa35a, 0xa167, 0x9f7f, 0x9da0, 0x9bc9, 0x99fa, 0x9831, 0x966e, 0x94b1, 0x92f9, 0x9145, 0x8f95, 0x8de8, 0x8c3e, 0x8a97, 0x88f2, 0x8750, 0x85ae, 0x840e, 0x826e, 0x80cf, 0x7f31, 0x7d92, 0x7bf2, 0x7a52, 0x78b0, 0x770e, 0x7569, 0x73c2, 0x7218, 0x706b, 0x6ebb, 0x6d07, 0x6b4f, 0x6992, 0x67cf, 0x6606, 0x6437, 0x6260, 0x6081, 0x5e99, 0x5ca6, 0x5aa9, 0x589e, 0x5686, 0x545e, 0x5224, 0x4fd5, 0x4d6f, 0x4aee, 0x484e, 0x4588, 0x4296, 0x3f6c, 0x3bfd, 0x3834, 0x33ee, 0x2ef3, 0x28cb, 0x202c, 0x1016};
 
-void Mycila::Dimmer::begin(int8_t pin, uint32_t semiPeriod) {
+void Mycila::Dimmer::begin(int8_t pin, uint16_t semiPeriod) {
   if (_enabled)
     return;
 
   if (!semiPeriod) {
-    LOGE(TAG, "Disable Dimmer on pin %" PRId8 ": Invalid semi-period: %" PRIu32 " us", pin, semiPeriod);
+    LOGE(TAG, "Disable Dimmer on pin %" PRId8 ": Invalid semi-period: %" PRIu16 " us", pin, semiPeriod);
     return;
   }
 
@@ -77,7 +77,7 @@ void Mycila::Dimmer::begin(int8_t pin, uint32_t semiPeriod) {
     return;
   }
 
-  LOGI(TAG, "Enable Dimmer on pin %" PRId8 " with semi-period %" PRIu32 " us", pin, semiPeriod);
+  LOGI(TAG, "Enable Dimmer on pin %" PRId8 " with semi-period %" PRIu16 " us", pin, semiPeriod);
 
   _semiPeriod = semiPeriod;
   pinMode(_pin, OUTPUT);
@@ -94,9 +94,8 @@ void Mycila::Dimmer::end() {
     return;
   _enabled = false;
   LOGI(TAG, "Disable Dimmer on pin %" PRId8, _pin);
-  // keep last saved value,
-  // _dutyCycle = 0;
-  _delay = UINT32_MAX;
+  // Note: do not set _dutyCycle to 0 in order to keep last set user value
+  _delay = UINT16_MAX;
   _semiPeriod = 0;
   _dimmer->setDelay(_delay);
   delete _dimmer;
@@ -145,7 +144,7 @@ void Mycila::Dimmer::setDutyCycleMax(float max) {
   setDutyCycle(_dutyCycle);
 }
 
-uint32_t Mycila::Dimmer::_lookupPhaseDelay(float dutyCycle) {
+uint16_t Mycila::Dimmer::_lookupPhaseDelay(float dutyCycle) {
   uint32_t duty = dutyCycle * DIMMER_MAX;
   uint32_t slot = duty * TABLE_PHASE_SCALE + (TABLE_PHASE_SCALE >> 1);
   uint32_t index = slot >> 16;
